requirements.txt
absl-py
astunparse
black
blinker
cachetools
certifi
charset-normalizer
click
colorama
Flask
Flask-Cors
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
idna
itsdangerous
Jinja2
keras
libclang
Markdown
MarkupSafe
ml-dtypes
mypy-extensions
numpy
oauthlib
opt-einsum
packaging
pathspec
platformdirs
protobuf
pyasn1
pyasn1-modules
python-dotenv
requests
requests-oauthlib
rsa
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-cpu
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
tomli
typing_extensions
urllib3
Werkzeug
wrapt
